rename core to cli
