+++ src/cli/reference_docs_generator.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 from typing import List
 
-from src.core.cli_app import CLIApp
-from src.core.command import Command
+from src.cli.cli_app import CLIApp
+from src.cli.command import Command
 
 
 class ReferenceDocsGenerator:
