+++ src/commands/remove/remove_command.py
@@ -5,7 +5,7 @@ from src.commands.install.install_command import (
     EXTERNAL_DEPENDENCY_REFERENCE_DESCRIPTION,
 )
 from src.commands.remove.remove_package import remove_package
-from src.core import Command
+from src.cli import Command
 from src.utils import Project, log_color_provider, retrieve_real_package_name
 
 INTERNAL_DEPENDENCY_REFERENCE_DESCRIPTION = (
