+++ src/commands/update/update_command.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 from src.commands.remove.remove_command import INTERNAL_DEPENDENCY_REFERENCE_DESCRIPTION
 from src.commands.update.update_package import update_package
 from src.commands.update.updating_exceptions import PackageAlreadyUpToDateException
-from src.core import Command
+from src.cli import Command
 from src.utils import Project, retrieve_real_package_name
