+++ src/conftest.py
@@ -2,7 +2,7 @@ from typing import List, Optional
 
 import pytest
 
-from src.core.command import Command
+from src.cli.command import Command
 
 
 class BaseTestCommand(Command):
